v1

+++ be-focus/resource-server/src/main/java/org/resourceserver/modules/stats/dto/StatsResponse.java
@@ -43,6 +43,6 @@ public class StatsResponse {
     @Builder
     public static class WeeklyData {
         private List<String> labels;
-        private List<Double> values;
+        private List<Integer> values;
     }
 }
